Extract FeedStoreSpy to shared file
