+++ EssentialFeed/EssentialFeedTests/FeedCache/LoadFeedFromCacheUseCase.swift
@@ -27,41 +27,3 @@ class LoadFeedFromCacheUseCase: XCTestCase {
     }
 
 }
-
-private  class FeedStoreSpy: FeedStore {
-    enum ReceivedMessage: Equatable {
-        case deleteCacheFeed
-        case insert([LocalFeedImage], Date)
-    }
-        
-    private var deletionCompletions: [DeletionCompletion] = []
-    private var insertionCompletions: [InsertionCompletion] = []
-    
-    private(set) var receivedMessages: [ReceivedMessage] = []
-    
-    func deleteCachedFeed(completion: @escaping DeletionCompletion) {
-        deletionCompletions.append(completion)
-        receivedMessages.append(.deleteCacheFeed)
-    }
-    
-    func completeDeletion(with error: Error, at index: Int = 0) {
-        deletionCompletions[index](error)
-    }
-    
-    func completeDeletionSuccessfully(at index: Int = 0) {
-        deletionCompletions[index](nil)
-    }
-    
-    func insert(_ feed: [LocalFeedImage], timestamp: Date, completion: @escaping InsertionCompletion) {
-        insertionCompletions.append(completion)
-        receivedMessages.append(.insert(feed, timestamp))
-    }
-    
-    func completeInsertion(with error: Error, at index: Int = 0) {
-        insertionCompletions[index](error)
-    }
-    
-    func completeInsertionSuccessfully(at index: Int = 0) {
-        insertionCompletions[index](nil)
-    }
-}
